Moved vega visualizations to geostreams package
